Create CheckForReadyRepo plug
Plug that does a couple of checks ror the repos of an `:otp_app`:

- If the storage is up / the database is created
- If there are pending migrations

Each of one raising an error if the condition is not met.

Adds `Ecto.Migrator.run/1` and `Ecto.Migrator.migrations/1` to xref
exclude config since we are not sure if `ecto_sql` dep will be loaded as
to not keep raising warnings about "undefined Ecto.Migrator module".